Convert slack to webpacked asset
