+++ client/slack/member_map.js
@@ -1,0 +1,417 @@
+import $ from 'jquery';
+
+import L from 'leaflet';
+
+const timezoneMapping = {
+    'Africa/Abidjan': ['5.316666', '-4.03333'],
+    'Africa/Accra': ['5.55', '-0.21666'],
+    'Africa/Algiers': ['36.78333', '3.05'],
+    'Africa/Bissau': ['11.85', '-15.5833'],
+    'Africa/Cairo': ['30.05', '31.25'],
+    'Africa/Casablanca': ['33.65', '-7.58333'],
+    'Africa/Ceuta': ['35.88333', '-5.31666'],
+    'Africa/El_Aaiun': ['27.15', '-13.2'],
+    'Africa/Harare': ['-17.824858', '31.053028'],
+    'Africa/Khartoum': ['15.6', '32.53333'],
+    'Africa/Lagos': ['6.45', '3.4'],
+    'Africa/Maputo': ['-25.9666', '32.58333'],
+    'Africa/Monrovia': ['6.3', '-10.7833'],
+    'Africa/Nairobi': ['-1.28333', '36.81666'],
+    'Africa/Ndjamena': ['12.11666', '15.05'],
+    'Africa/Tripoli': ['32.9', '13.18333'],
+    'Africa/Tunis': ['36.8', '10.18333'],
+    'Africa/Windhoek': ['-22.5666', '17.1'],
+    'America/Adak': ['51.88000', '-176.658'],
+    'America/Anchorage': ['61.21805', '-149.900'],
+    'America/Araguaina': ['-7.2', '-48.2'],
+    'America/Buenos_Aires': ['-34.6', '-58.45'],
+    'America/Argentina/Buenos_Aires': ['-34.6', '-58.45'],
+    'America/Argentina/Catamarca': ['-28.4666', '-65.7833'],
+    'America/Argentina/Cordoba': ['-31.4', '-64.1833'],
+    'America/Argentina/Jujuy': ['-24.1833', '-65.3'],
+    'America/Argentina/La_Rioja': ['-29.4333', '-66.85'],
+    'America/Argentina/Mendoza': ['-32.8833', '-68.8166'],
+    'America/Argentina/Rio_Gallegos': ['-51.6333', '-69.2166'],
+    'America/Argentina/Salta': ['-24.7833', '-65.4166'],
+    'America/Argentina/San_Juan': ['-31.5333', '-68.5166'],
+    'America/Argentina/San_Luis': ['-33.3166', '-66.35'],
+    'America/Argentina/Tucuman': ['-26.8166', '-65.2166'],
+    'America/Argentina/Ushuaia': ['-54.8', '-68.3'],
+    'America/Asuncion': ['-25.2666', '-57.6666'],
+    'America/Atikokan': ['48.75861', '-91.6216'],
+    'America/Bahia': ['-12.9833', '-38.5166'],
+    'America/Bahia_Banderas': ['20.8', '-105.25'],
+    'America/Barbados': ['13.1', '-59.6166'],
+    'America/Belem': ['-1.45', '-48.4833'],
+    'America/Belize': ['17.5', '-88.2'],
+    'America/Blanc-Sablon': ['51.41666', '-57.1166'],
+    'America/Boa_Vista': ['2.816666', '-60.6666'],
+    'America/Bogota': ['4.6', '-74.0833'],
+    'America/Boise': ['43.61361', '-116.202'],
+    'America/Cambridge_Bay': ['69.11388', '-105.052'],
+    'America/Campo_Grande': ['-20.45', '-54.6166'],
+    'America/Cancun': ['21.08333', '-86.7666'],
+    'America/Caracas': ['10.5', '-66.9333'],
+    'America/Cayenne': ['4.933333', '-52.3333'],
+    'America/Chicago': ['41.85', '-87.65'],
+    'America/Chihuahua': ['28.63333', '-106.083'],
+    'America/Costa_Rica': ['9.933333', '-84.0833'],
+    'America/Creston': ['49.1', '-116.516'],
+    'America/Cuiaba': ['-15.5833', '-56.0833'],
+    'America/Curacao': ['12.18333', '-69'],
+    'America/Danmarkshavn': ['76.76666', '-18.6666'],
+    'America/Dawson': ['64.06666', '-139.416'],
+    'America/Dawson_Creek': ['59.76666', '-120.233'],
+    'America/Denver': ['39.73916', '-104.984'],
+    'America/Detroit': ['42.33138', '-83.0458'],
+    'America/Edmonton': ['53.55', '-113.466'],
+    'America/Eirunepe': ['-6.66666', '-69.8666'],
+    'America/El_Salvador': ['13.7', '-89.2'],
+    'America/Fort_Nelson': ['58.8', '-122.7'],
+    'America/Fortaleza': ['-3.71666', '-38.5'],
+    'America/Glace_Bay': ['46.2', '-59.95'],
+    'America/Godthab': ['64.18333', '-51.7333'],
+    'America/Goose_Bay': ['53.33333', '-60.4166'],
+    'America/Grand_Turk': ['21.46666', '-71.1333'],
+    'America/Guatemala': ['14.63333', '-90.5166'],
+    'America/Guayaquil': ['-2.16666', '-79.8333'],
+    'America/Guyana': ['6.8', '-58.1666'],
+    'America/Halifax': ['44.65', '-63.6'],
+    'America/Havana': ['23.13333', '-82.3666'],
+    'America/Hermosillo': ['29.06666', '-110.966'],
+    'America/Indiana/Indianapolis': ['39.76833', '-86.1580'],
+    'America/Indiana/Knox': ['41.29583', '-86.6250'],
+    'America/Indiana/Marengo': ['38.37555', '-86.3447'],
+    'America/Indiana/Petersburg': ['38.49194', '-87.2786'],
+    'America/Indiana/Tell_City': ['37.95305', '-86.7613'],
+    'America/Indiana/Vevay': ['38.74777', '-85.0672'],
+    'America/Indiana/Vincennes': ['38.67722', '-87.5286'],
+    'America/Indiana/Winamac': ['41.05138', '-86.6030'],
+    'America/Inuvik': ['68.34972', '-133.716'],
+    'America/Iqaluit': ['63.73333', '-68.4666'],
+    'America/Jamaica': ['17.96805', '-76.7933'],
+    'America/Juneau': ['58.30194', '-134.419'],
+    'America/Kentucky/Louisville': ['38.25416', '-85.7594'],
+    'America/Kentucky/Monticello': ['36.82972', '-84.8491'],
+    'America/La_Paz': ['-16.5', '-68.15'],
+    'America/Lima': ['-12.05', '-77.05'],
+    'America/Los_Angeles': ['34.05222', '-118.242'],
+    'America/Maceio': ['-9.66666', '-35.7166'],
+    'America/Managua': ['12.15', '-86.2833'],
+    'America/Manaus': ['-3.13333', '-60.0166'],
+    'America/Martinique': ['14.6', '-61.0833'],
+    'America/Matamoros': ['25.83333', '-97.5'],
+    'America/Mazatlan': ['23.21666', '-106.416'],
+    'America/Menominee': ['45.10777', '-87.6141'],
+    'America/Merida': ['20.96666', '-89.6166'],
+    'America/Metlakatla': ['55.12694', '-131.576'],
+    'America/Mexico_City': ['19.4', '-99.15'],
+    'America/Miquelon': ['47.05', '-56.3333'],
+    'America/Moncton': ['46.1', '-64.7833'],
+    'America/Monterrey': ['25.66666', '-100.316'],
+    'America/Montevideo': ['-34.8833', '-56.1833'],
+    'America/Nassau': ['25.08333', '-77.35'],
+    'America/New_York': ['40.71416', '-74.0063'],
+    'America/Nipigon': ['49.01666', '-88.2666'],
+    'America/Nome': ['64.50111', '-165.406'],
+    'America/Noronha': ['-3.85', '-32.4166'],
+    'America/North_Dakota/Beulah': ['47.26416', '-101.777'],
+    'America/North_Dakota/Center': ['47.11638', '-101.299'],
+    'America/North_Dakota/New_Salem': ['46.84500', '-101.410'],
+    'America/Ojinaga': ['29.56666', '-104.416'],
+    'America/Panama': ['8.966666', '-79.5333'],
+    'America/Pangnirtung': ['66.13333', '-65.7333'],
+    'America/Paramaribo': ['5.833333', '-55.1666'],
+    'America/Phoenix': ['33.44833', '-112.073'],
+    'America/Port-au-Prince': ['18.53333', '-72.3333'],
+    'America/Port_of_Spain': ['10.65', '-61.5166'],
+    'America/Porto_Velho': ['-8.76666', '-63.9'],
+    'America/Puerto_Rico': ['18.46833', '-66.1061'],
+    'America/Punta_Arenas': ['-53.15', '-70.9166'],
+    'America/Rainy_River': ['48.71666', '-94.5666'],
+    'America/Rankin_Inlet': ['62.81666', '-92.0830'],
+    'America/Recife': ['-8.05', '-34.9'],
+    'America/Regina': ['50.4', '-104.65'],
+    'America/Resolute': ['74.69555', '-94.8291'],
+    'America/Rio_Branco': ['-9.96666', '-67.8'],
+    'America/Santarem': ['-2.43333', '-54.8666'],
+    'America/Santiago': ['-33.45', '-70.6666'],
+    'America/Santo_Domingo': ['18.46666', '-69.9'],
+    'America/Sao_Paulo': ['-23.5333', '-46.6166'],
+    'America/Scoresbysund': ['70.48333', '-21.9666'],
+    'America/Sitka': ['57.17638', '-135.301'],
+    'America/St_Johns': ['47.56666', '-52.7166'],
+    'America/Swift_Current': ['50.28333', '-107.833'],
+    'America/Tegucigalpa': ['14.1', '-87.2166'],
+    'America/Thule': ['76.56666', '-68.7833'],
+    'America/Thunder_Bay': ['48.38333', '-89.25'],
+    'America/Tijuana': ['32.53333', '-117.016'],
+    'America/Toronto': ['43.65', '-79.3833'],
+    'America/Vancouver': ['49.26666', '-123.116'],
+    'America/Whitehorse': ['60.71666', '-135.05'],
+    'America/Winnipeg': ['49.88333', '-97.15'],
+    'America/Yakutat': ['59.54694', '-139.727'],
+    'America/Yellowknife': ['62.45', '-114.35'],
+    'Antarctica/Casey': ['-66.2833', '110.5166'],
+    'Antarctica/Davis': ['-68.5833', '77.96666'],
+    'Antarctica/DumontDUrville': ['-66.6666', '140.0166'],
+    'Antarctica/Macquarie': ['-54.5', '158.95'],
+    'Antarctica/Mawson': ['-67.6', '62.88333'],
+    'Antarctica/Palmer': ['-64.8', '-64.1'],
+    'Antarctica/Rothera': ['-67.5666', '-68.1333'],
+    'Antarctica/Syowa': ['-69.0061', '39.59000'],
+    'Antarctica/Troll': ['-72.0113', '2.535000'],
+    'Antarctica/Vostok': ['-78.4', '106.9'],
+    'Asia/Almaty': ['43.25', '76.95'],
+    'Asia/Amman': ['31.95', '35.93333'],
+    'Asia/Anadyr': ['64.75', '177.4833'],
+    'Asia/Aqtau': ['44.51666', '50.26666'],
+    'Asia/Aqtobe': ['50.28333', '57.16666'],
+    'Asia/Ashgabat': ['37.95', '58.38333'],
+    'Asia/Atyrau': ['47.11666', '51.93333'],
+    'Asia/Baghdad': ['33.35', '44.41666'],
+    'Asia/Baku': ['40.38333', '49.85'],
+    'Asia/Bangkok': ['13.75', '100.5166'],
+    'Asia/Barnaul': ['53.36666', '83.75'],
+    'Asia/Beirut': ['33.88333', '35.5'],
+    'Asia/Bishkek': ['42.9', '74.6'],
+    'Asia/Brunei': ['4.933333', '114.9166'],
+    'Asia/Chita': ['52.05', '113.4666'],
+    'Asia/Choibalsan': ['48.06666', '114.5'],
+    'Asia/Colombo': ['6.933333', '79.85'],
+    'Asia/Damascus': ['33.5', '36.3'],
+    'Asia/Dhaka': ['23.71666', '90.41666'],
+    'Asia/Dili': ['-8.55', '125.5833'],
+    'Asia/Dubai': ['25.3', '55.3'],
+    'Asia/Dushanbe': ['38.58333', '68.8'],
+    'Asia/Famagusta': ['35.11666', '33.95'],
+    'Asia/Gaza': ['31.5', '34.46666'],
+    'Asia/Hebron': ['31.53333', '35.09500'],
+    'Asia/Ho_Chi_Minh': ['10.75', '106.6666'],
+    'Asia/Hong_Kong': ['22.28333', '114.15'],
+    'Asia/Hovd': ['48.01666', '91.65'],
+    'Asia/Irkutsk': ['52.26666', '104.3333'],
+    'Asia/Jakarta': ['-6.16666', '106.8'],
+    'Asia/Jayapura': ['-2.53333', '140.7'],
+    'Asia/Jerusalem': ['31.78055', '35.22388'],
+    'Asia/Kabul': ['34.51666', '69.2'],
+    'Asia/Kamchatka': ['53.01666', '158.65'],
+    'Asia/Karachi': ['24.86666', '67.05'],
+    'Asia/Kathmandu': ['27.71666', '85.31666'],
+    'Asia/Katmandu': ['27.71666', '85.31666'],
+    'Asia/Khandyga': ['62.65638', '135.5538'],
+    'Asia/Kolkata': ['22.53333', '88.36666'],
+    'Asia/Krasnoyarsk': ['56.01666', '92.83333'],
+    'Asia/Kuala_Lumpur': ['3.166666', '101.7'],
+    'Asia/Kuching': ['1.55', '110.3333'],
+    'Asia/Kuwait': ['29.378586', '47.990341'],
+    'Asia/Macau': ['22.23333', '113.5833'],
+    'Asia/Magadan': ['59.56666', '150.8'],
+    'Asia/Makassar': ['-5.11666', '119.4'],
+    'Asia/Manila': ['14.58333', '121'],
+    'Asia/Muscat': ['23.614328', '58.545284'],
+    'Asia/Nicosia': ['35.16666', '33.36666'],
+    'Asia/Novokuznetsk': ['53.75', '87.11666'],
+    'Asia/Novosibirsk': ['55.03333', '82.91666'],
+    'Asia/Omsk': ['55', '73.4'],
+    'Asia/Oral': ['51.21666', '51.35'],
+    'Asia/Pontianak': ['-0.03333', '109.3333'],
+    'Asia/Pyongyang': ['39.01666', '125.75'],
+    'Asia/Qatar': ['25.28333', '51.53333'],
+    'Asia/Qyzylorda': ['44.8', '65.46666'],
+    'Asia/Riyadh': ['24.63333', '46.71666'],
+    'Asia/Sakhalin': ['46.96666', '142.7'],
+    'Asia/Samarkand': ['39.66666', '66.8'],
+    'Asia/Seoul': ['37.55', '126.9666'],
+    'Asia/Shanghai': ['31.23333', '121.4666'],
+    'Asia/Chongqing': ['29.56278', '106.55278'],
+    'Asia/Singapore': ['1.283333', '103.85'],
+    'Asia/Srednekolymsk': ['67.46666', '153.7166'],
+    'Asia/Taipei': ['25.05', '121.5'],
+    'Asia/Tashkent': ['41.33333', '69.3'],
+    'Asia/Tbilisi': ['41.71666', '44.81666'],
+    'Asia/Tehran': ['35.66666', '51.43333'],
+    'Asia/Thimphu': ['27.46666', '89.65'],
+    'Asia/Tokyo': ['35.65444', '139.7447'],
+    'Asia/Tomsk': ['56.5', '84.96666'],
+    'Asia/Ulaanbaatar': ['47.91666', '106.8833'],
+    'Asia/Urumqi': ['43.8', '87.58333'],
+    'Asia/Ust-Nera': ['64.56027', '143.2266'],
+    'Asia/Vladivostok': ['43.16666', '131.9333'],
+    'Asia/Yakutsk': ['62', '129.6666'],
+    'Asia/Yangon': ['16.78333', '96.16666'],
+    'Asia/Rangoon': ['16.78333', '96.16666'],
+    'Asia/Yekaterinburg': ['56.85', '60.6'],
+    'Asia/Yerevan': ['40.18333', '44.5'],
+    'Atlantic/Azores': ['37.73333', '-25.6666'],
+    'Atlantic/Bermuda': ['32.28333', '-64.7666'],
+    'Atlantic/Canary': ['28.1', '-15.4'],
+    'Atlantic/Cape_Verde': ['14.91666', '-23.5166'],
+    'Atlantic/Faroe': ['62.01666', '-6.76666'],
+    'Atlantic/Madeira': ['32.63333', '-16.9'],
+    'Atlantic/Reykjavik': ['64.15', '-21.85'],
+    'Atlantic/South_Georgia': ['-54.2666', '-36.5333'],
+    'Atlantic/Stanley': ['-51.7', '-57.85'],
+    'Australia/Adelaide': ['-34.9166', '138.5833'],
+    'Australia/Brisbane': ['-27.4666', '153.0333'],
+    'Australia/Broken_Hill': ['-31.95', '141.45'],
+    'Australia/Canberra': ['-35.343784', '149.082977'],
+    'Australia/Currie': ['-39.9333', '143.8666'],
+    'Australia/Darwin': ['-12.4666', '130.8333'],
+    'Australia/Eucla': ['-31.7166', '128.8666'],
+    'Australia/Hobart': ['-42.8833', '147.3166'],
+    'Australia/Lindeman': ['-20.2666', '149'],
+    'Australia/Lord_Howe': ['-31.55', '159.0833'],
+    'Australia/Melbourne': ['-37.8166', '144.9666'],
+    'Australia/Perth': ['-31.95', '115.85'],
+    'Australia/Sydney': ['-33.8666', '151.2166'],
+    'Europe/Amsterdam': ['52.36666', '4.9'],
+    'Europe/Andorra': ['42.5', '1.516666'],
+    'Europe/Astrakhan': ['46.35', '48.05'],
+    'Europe/Athens': ['37.96666', '23.71666'],
+    'Europe/Belgrade': ['44.83333', '20.5'],
+    'Europe/Berlin': ['52.5', '13.36666'],
+    'Europe/Brussels': ['50.83333', '4.333333'],
+    'Europe/Bucharest': ['44.43333', '26.1'],
+    'Europe/Budapest': ['47.5', '19.08333'],
+    'Europe/Chisinau': ['47', '28.83333'],
+    'Europe/Copenhagen': ['55.66666', '12.58333'],
+    'Europe/Dublin': ['53.33333', '-6.25'],
+    'Europe/Gibraltar': ['36.13333', '-5.35'],
+    'Europe/Helsinki': ['60.16666', '24.96666'],
+    'Asia/Istanbul': ['41.01666', '28.96666'],
+    'Europe/Istanbul': ['41.01666', '28.96666'],
+    'Europe/Kaliningrad': ['54.71666', '20.5'],
+    'Europe/Kiev': ['50.43333', '30.51666'],
+    'Europe/Kirov': ['58.6', '49.65'],
+    'Europe/Lisbon': ['38.71666', '-9.13333'],
+    'Europe/London': ['51.50833', '-0.12527'],
+    'Europe/Luxembourg': ['49.6', '6.15'],
+    'Europe/Madrid': ['40.4', '-3.68333'],
+    'Europe/Malta': ['35.9', '14.51666'],
+    'Europe/Minsk': ['53.9', '27.56666'],
+    'Europe/Monaco': ['43.7', '7.383333'],
+    'Europe/Moscow': ['55.75583', '37.61777'],
+    'Europe/Oslo': ['59.91666', '10.75'],
+    'Europe/Paris': ['48.86666', '2.333333'],
+    'Europe/Prague': ['50.08333', '14.43333'],
+    'Europe/Riga': ['56.95', '24.1'],
+    'Europe/Rome': ['41.9', '12.48333'],
+    'Europe/Samara': ['53.2', '50.15'],
+    'Europe/Saratov': ['51.56666', '46.03333'],
+    'Europe/Simferopol': ['44.95', '34.1'],
+    'Europe/Sofia': ['42.68333', '23.31666'],
+    'Europe/Stockholm': ['59.33333', '18.05'],
+    'Europe/Tallinn': ['59.41666', '24.75'],
+    'Europe/Tirane': ['41.33333', '19.83333'],
+    'Europe/Ulyanovsk': ['54.33333', '48.4'],
+    'Europe/Uzhgorod': ['48.61666', '22.3'],
+    'Europe/Vienna': ['48.21666', '16.33333'],
+    'Europe/Vilnius': ['54.68333', '25.31666'],
+    'Europe/Volgograd': ['48.73333', '44.41666'],
+    'Europe/Warsaw': ['52.25', '21'],
+    'Europe/Zaporozhye': ['47.83333', '35.16666'],
+    'Europe/Zurich': ['47.38333', '8.533333'],
+    'Indian/Chagos': ['-7.33333', '72.41666'],
+    'Indian/Christmas': ['-10.4166', '105.7166'],
+    'Indian/Cocos': ['-12.1666', '96.91666'],
+    'Indian/Kerguelen': ['-49.3527', '70.21750'],
+    'Indian/Mahe': ['-4.66666', '55.46666'],
+    'Indian/Maldives': ['4.166666', '73.5'],
+    'Indian/Mauritius': ['-20.1666', '57.5'],
+    'Indian/Reunion': ['-20.8666', '55.46666'],
+    'Pacific/Apia': ['-13.8333', '-171.733'],
+    'Pacific/Auckland': ['-36.8666', '174.7666'],
+    'Pacific/Bougainville': ['-6.21666', '155.5666'],
+    'Pacific/Chatham': ['-43.95', '-176.55'],
+    'Pacific/Chuuk': ['7.416666', '151.7833'],
+    'Pacific/Easter': ['-27.15', '-109.433'],
+    'Pacific/Efate': ['-17.6666', '168.4166'],
+    'Pacific/Enderbury': ['-3.13333', '-171.083'],
+    'Pacific/Fakaofo': ['-9.36666', '-171.233'],
+    'Pacific/Fiji': ['-18.1333', '178.4166'],
+    'Pacific/Funafuti': ['-8.51666', '179.2166'],
+    'Pacific/Galapagos': ['-0.9', '-89.6'],
+    'Pacific/Gambier': ['-23.1333', '-134.95'],
+    'Pacific/Guadalcanal': ['-9.53333', '160.2'],
+    'Pacific/Guam': ['13.46666', '144.75'],
+    'Pacific/Honolulu': ['21.30694', '-157.858'],
+    'Pacific/Kiritimati': ['1.866666', '-157.333'],
+    'Pacific/Kosrae': ['5.316666', '162.9833'],
+    'Pacific/Kwajalein': ['9.083333', '167.3333'],
+    'Pacific/Majuro': ['7.15', '171.2'],
+    'Pacific/Marquesas': ['-9', '-139.5'],
+    'Pacific/Midway': ['28.1999992', '-177.3499986'],
+    'Pacific/Nauru': ['-0.51666', '166.9166'],
+    'Pacific/Niue': ['-19.0166', '-169.916'],
+    'Pacific/Norfolk': ['-29.05', '167.9666'],
+    'Pacific/Noumea': ['-22.2666', '166.45'],
+    'Pacific/Pago_Pago': ['-14.2666', '-170.7'],
+    'Pacific/Palau': ['7.333333', '134.4833'],
+    'Pacific/Pitcairn': ['-25.0666', '-130.083'],
+    'Pacific/Pohnpei': ['6.966666', '158.2166'],
+    'Pacific/Port_Moresby': ['-9.5', '147.1666'],
+    'Pacific/Rarotonga': ['-21.2333', '-159.766'],
+    'Pacific/Tahiti': ['-17.5333', '-149.566'],
+    'Pacific/Tarawa': ['1.416666', '173'],
+    'Pacific/Tongatapu': ['-21.1666', '-175.166'],
+    'Pacific/Wake': ['19.28333', '166.6166'],
+    'Pacific/Wallis': ['-13.3', '-176.166']
+};
+
+function buildMap(slackMemberTZCount) {
+    const features = slackMemberTZCount.map(([timezone, memberCount]) => ({
+        geometry: {
+            type: 'Point',
+            coordinates: [
+                timezoneMapping[timezone][1],
+                timezoneMapping[timezone][0]
+            ]
+        },
+        type: 'Feature',
+        properties: {
+            popupContent: `<h3>${timezone}: ${memberCount} members</h3>`,
+            member_count: memberCount
+        },
+    }));
+
+    const map = L.map('leaflet_member_map', {
+        minZoom: 1.5,
+        maxZoom: 5
+    });
+
+    L.tileLayer('https://cartocdn_{s}.global.ssl.fastly.net/base-light/{z}/{x}/{y}.png', {
+        attribution: 'PySlackers Member List by Time Zone | Maps by cartocdn.com',
+        noWrap: true
+    }).addTo(map);
+
+    L.geoJSON(features, {
+        onEachFeature: (feature, layer) => layer.bindPopup(feature.properties.popupContent),
+        pointToLayer: (feature, latlng) => L.circleMarker(latlng, {
+            radius: Math.log10(feature.properties.member_count) * 4,
+            fillColor: '#3c7da7',
+            color: '#000',
+            weight: 1,
+            opacity: 1,
+            fillOpacity: 0.8
+        })
+    }).addTo(map);
+
+    map.setView([10, 0], 0);
+    return map;
+}
+
+export default (slackMemberTimezones) => {
+    let map;
+    $(window).on('orientationchange pageshow resize', () => {
+        if (!map) {
+            map = buildMap(slackMemberTimezones);
+        }
+        const mapElem = $('#leaflet_member_map');
+        const mapContainer = $('#leaflet_container');
+        mapElem.height(mapContainer.height());
+        mapElem.width(mapContainer.width());
+        map.invalidateSize();
+        map.setView([10, 0], 0);
+    }).trigger('resize');
+};
