+++ client/style/app.scss
@@ -1,4 +1,5 @@
 @import "~semantic-ui-css/semantic.css";
+@import "~leaflet/dist/leaflet.css";
 @import 'codehilite';
